backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
slowapi==0.1.9
redis==5.0.1
python-dotenv==1.0.0
networkx==3.1
pyvis==0.3.2
reportlab==4.0.4
Pillow==10.0.1
pydantic==2.4.2